# swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Authentication API",
    "description": "API for user authentication, including login, password reset request, OTP verification, and password reset.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "paths": {
...
    "/classes": {
      "get": {
        "summary": "Retrieve Class",
        "description": "Retrieves a Class resource",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Class"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/boards": {
      "get": {
        "summary": "Retrieve Board",
        "description": "Retrieves a Board resource",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/student/presigned-url": {
      "post": {
        "summary": "Create Student",
        "description": "Creates a Student resource",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        }
      }
    },
    "/createstudent": {
      "put": {
        "summary": "Update Student",
        "description": "Updates a Student resource",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        }
      }
    },
    "/student/:id": {
      "put": {
        "summary": "Update Student",
        "description": "Updates a Student resource",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve Student",
        "description": "Retrieves a Student resource",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/presigned-url": {
      "get": {
        "summary": "Retrieve presigned url Resource",
        "description": "Retrieves a resource at /presigned-url",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create presigned url Resource",
        "description": "Creates a resource at /presigned-url",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Board": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "language": {
            "type": "objectid"
          },
          "_id": {
            "type": "string",
            "description": "MongoDB ObjectId",
            "example": "507f1f77bcf86cd799439011"
          }
        }
      },
...
      "Class": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "_id": {
            "type": "string",
            "description": "MongoDB ObjectId",
            "example": "507f1f77bcf86cd799439011"
          }
        }
      },
...
      "Student": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "objectid"
          },
          "isApproved": {
            "type": "boolean"
          },
          "rejected": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "studentUid": {
            "type": "string"
          },
          "uid": {
            "type": "string"
          },
          "referralCode": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "pinCode": {
            "type": "string"
          },
          "parentsMobile": {
            "type": "string"
          },
          "parentEmail": {
            "type": "string"
          },
          "fatherName": {
            "type": "string"
          },
          "motherName": {
            "type": "string"
          },
          "fatherOccupation": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "fatherTitle": {
            "type": "string"
          },
          "motherTitle": {
            "type": "string"
          },
          "motherOccupation": {
            "type": "string"
          },
          "classId": {
            "type": "objectid"
          },
          "className": {
            "type": "string"
          },
          "termsAccepted": {
            "type": "boolean"
          },
          "schoolName": {
            "type": "string"
          },
          "boardId": {
            "type": "objectid"
          },
          "boardName": {
            "type": "string"
          },
          "billingCode": {
            "type": "string"
          },
          "affiliatedReferralCode": {
            "type": "string"
          },
          "taggedTeacher": {
            "type": "objectid"
          },
          "avatarUrl": {
            "type": "string"
          },
          "assignedTeachers": {
            "type": "array"
          },
          "isBlocked": {
            "type": "boolean"
          },
          "createdByAdmin": {
            "type": "boolean"
          },
          "_id": {
            "type": "string",
            "description": "MongoDB ObjectId",
            "example": "507f1f77bcf86cd799439011"
          },
          "createdAt": {
            "type": "date"
          },
          "updatedAt": {
            "type": "date"
          }
        }
      },
...
    }
  }
}